refactor: migrate organization-related imports to organization-utils module
