+++ packages/insomnia/src/ui/components/modals/upgrade-modal.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef, useImperativeHandle, useRef } from 'react';
 
 import { getAppWebsiteBaseURL } from '../../../common/constants';
-import type { PersonalPlanType } from '../../../models/organization';
+import type { PersonalPlanType } from '../../organization-utils';
 import { type ModalProps } from '../base/modal';
 import { AskModal, type AskModalHandle, type AskModalOptions } from './ask-modal';
