+++ packages/insomnia/src/ui/components/modals/workspace-settings-modal.tsx
@@ -19,8 +19,8 @@ import { database as db } from '../../../common/database';
 import { getWorkspaceLabel } from '../../../common/get-workspace-label';
 import * as models from '../../../models/index';
 import type { MockServer } from '../../../models/mock-server';
-import { fetchAndCacheOrganizationStorageRule } from '../../../models/organization';
-import { DEFAULT_STORAGE_RULES, type StorageRules } from '../../../models/organization';
+import { fetchAndCacheOrganizationStorageRule } from '../../organization-utils';
+import { DEFAULT_STORAGE_RULES, type StorageRules } from '../../organization-utils';
 import { isGitProject, type Project } from '../../../models/project';
 import { isRequest } from '../../../models/request';
 import { isEnvironment, isMockServer, isScratchpad, type Workspace } from '../../../models/workspace';
