+++ packages/insomnia/src/ui/routes/organization.sync.tsx
@@ -1,7 +1,7 @@
 import type { ActionFunctionArgs } from 'react-router';
 
 import { userSession } from '../../models';
-import { syncOrganizations } from '../../models/organization';
+import { syncOrganizations } from '../organization-utils';
 
 export async function action(_args: ActionFunctionArgs) {
   const { id: sessionId, accountId } = await userSession.getOrCreate();
